refactor: extract area drawing hook
